+++ src/drivers/integrations/FeedMeIntegration.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\blitz\drivers\integrations;
+
+use craft\feedme\services\Process;
+use putyourlightson\blitz\Blitz;
+use yii\base\Event;
+
+class FeedMeIntegration extends BaseIntegration
+{
+    // Static Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public static function getRequiredPluginHandles(): array
+    {
+        return ['feed-me'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function registerEvents()
+    {
+        Event::on(Process::class, Process::EVENT_BEFORE_PROCESS_FEED,
+            function() {
+                Blitz::$plugin->refreshCache->batchMode = true;
+            }
+        );
+
+        Event::on(Process::class, Process::EVENT_AFTER_PROCESS_FEED,
+            function() {
+                Blitz::$plugin->refreshCache->refresh();
+            }
+        );
+    }
+}