+++ src/jobs/RefreshCacheJob.php
@@ -8,7 +8,6 @@ namespace putyourlightson\blitz\jobs;
 use Craft;
 use craft\base\Element;
 use craft\elements\db\ElementQuery;
-use craft\helpers\App;
 use craft\helpers\Json;
 use craft\queue\BaseJob;
 use Exception;
@@ -52,61 +51,70 @@ class RefreshCacheJob extends BaseJob
      */
     public function execute($queue)
     {
-        App::maxPowerCaptain();
-
-        // Step 1
-        $this->setProgress($queue, 1/3);
-
-        // Merge in element cache IDs
-        $this->cacheIds = array_merge($this->cacheIds,
-            Blitz::$plugin->refreshCache->getElementCacheIds(
-                $this->elementIds, $this->cacheIds
-            )
+        // Set progress label
+        $this->setProgress($queue, 0,
+            Craft::t('blitz', 'Finding cached pages.')
         );
 
-        if (!empty($this->elementIds)) {
+        // Merge in element cache IDs
+        $elementCacheIds = Blitz::$plugin->refreshCache->getElementCacheIds(
+            $this->elementIds, $this->cacheIds
+        );
+
+        $cacheIds = array_merge($this->cacheIds, $elementCacheIds);
+
+        if (count($this->elementIds)) {
             $elementQueryRecords = Blitz::$plugin->refreshCache->getElementTypeQueries(
-                $this->elementTypes, $this->cacheIds
+                $this->elementTypes, $cacheIds
             );
 
-            // Use sets and the splat operator rather than array_merge for performance (https://goo.gl/9mntEV)
-            $elementQueryCacheIdSets = [[]];
+            if (count($elementQueryRecords)) {
+                // Set progress total to number of query records plus one to avoid dividing by zero
+                $total = count($elementQueryRecords) + 1;
+                $count = 0;
 
-            foreach ($elementQueryRecords as $elementQueryRecord) {
-                // Merge in element query cache IDs
-                $elementQueryCacheIdSets[] = $this->_getElementQueryCacheIds(
-                    $elementQueryRecord, $this->elementIds, $this->cacheIds
-                );
+                // Use sets and the splat operator rather than array_merge for performance (https://goo.gl/9mntEV)
+                $elementQueryCacheIdSets = [[]];
+
+                foreach ($elementQueryRecords as $elementQueryRecord) {
+                    // Merge in element query cache IDs
+                    $elementQueryCacheIdSets[] = $this->_getElementQueryCacheIds(
+                        $elementQueryRecord, $this->elementIds, $cacheIds
+                    );
+
+                    $count++;
+                    $this->setProgress($queue, $count / $total,
+                        Craft::t('blitz', 'Checked {count} of {total} element queries.', [
+                            'count' => $count,
+                            'total' => $total,
+                        ])
+                    );
+                }
+
+                $elementQueryCacheIds = array_merge(...$elementQueryCacheIdSets);
+                $cacheIds = array_merge($cacheIds, $elementQueryCacheIds);
             }
-
-            $elementQueryCacheIds = array_merge(...$elementQueryCacheIdSets);
-
-            $this->cacheIds = array_merge($this->cacheIds, $elementQueryCacheIds);
         }
 
-        if (empty($this->cacheIds)) {
+        if (empty($cacheIds)) {
             return;
         }
 
-        // Step 2
-        $this->setProgress($queue, 2/3);
-
         // If clear automatically is enabled or if force clear
         if (Blitz::$plugin->settings->clearCacheAutomatically || $this->forceClear) {
-            // Get cached site URIs from cache IDs
-            $siteUris = SiteUriHelper::getCachedSiteUris($this->cacheIds);
+            // Set progress label
+            $this->setProgress($queue, 1,
+                Craft::t('blitz', 'Clearing {total} cached pages.', [
+                    'total' => count($cacheIds)
+                ])
+            );
 
-            // Delete cache records so we get fresh caches
-            Blitz::$plugin->flushCache->flushCacheIds($this->cacheIds);
+            $siteUris = SiteUriHelper::getCachedSiteUris($cacheIds);
 
-            // Delete cached values so we get a fresh version
-            Blitz::$plugin->cacheStorage->deleteUris($siteUris);
-
-            // Trigger afterRefreshCache events
-            Blitz::$plugin->refreshCache->afterRefresh($siteUris);
+            Blitz::$plugin->refreshCache->refreshSiteUris($siteUris);
         }
         else {
-            Blitz::$plugin->refreshCache->expireCacheIds($this->cacheIds);
+            Blitz::$plugin->refreshCache->expireCacheIds($cacheIds);
         }
     }