+++ src/migrations/Install.php
@@ -9,6 +9,7 @@ use Craft;
 use craft\db\Migration;
 use craft\records\Element;
 use craft\records\Site;
+use putyourlightson\blitz\models\SiteUriModel;
 use putyourlightson\blitz\records\CacheRecord;
 use putyourlightson\blitz\records\ElementCacheRecord;
 use putyourlightson\blitz\records\ElementExpiryDateRecord;
@@ -44,9 +45,6 @@ class Install extends Migration
      */
     public function safeDown(): bool
     {
-        // TODO: Remove in 2.1.0
-        $this->dropTableIfExists('{{%blitz_cacheflags}}');
-
         $this->dropTableIfExists(ElementQueryCacheRecord::tableName());
         $this->dropTableIfExists(ElementQueryRecord::tableName());
         $this->dropTableIfExists(ElementCacheRecord::tableName());
@@ -71,7 +69,7 @@ class Install extends Migration
             $this->createTable(CacheRecord::tableName(), [
                 'id' => $this->primaryKey(),
                 'siteId' => $this->integer()->notNull(),
-                'uri' => $this->string()->notNull(),
+                'uri' => $this->string(SiteUriModel::MAX_URI_LENGTH)->notNull(),
                 'expiryDate' => $this->dateTime(),
             ]);
         }